requirements.txt
Flask
pytest
supabase
python-dotenv
